Gestion des erreurs des Endpoints du module Apprenant

+++ src/models/apprenants/routes/deleteApprenant.js
@@ -6,23 +6,25 @@ module.exports = (app) => {
     app.delete('/api/udemy/apprenants/:id', (req, res) => {
 
         const id = req.params.id
-        Apprenant.findByPk(id)
-            .then(apprenant => {
-                const apprenantDeleted = apprenant
-                
-                Apprenant.destroy({
-                    where: { id: id }
-                })
-                .then(_ => {
-                    res.json({ message: `Apprenant \' ${apprenantDeleted.fullName} \' Deleted`, data: apprenantDeleted })
+        return  Apprenant.findByPk(id)
+                .then(apprenant => {
+                    if(apprenant === null){
+                        const message = `L'apprenant avec ID ${req.params.id} n'existe pas...`
+                        return res.status(404).json({ message })
+                    }
+                    const apprenantDeleted = apprenant
+                    
+                    Apprenant.destroy({
+                        where: { id: id }
+                    })
+                    .then(_ => {
+                        res.json({ message: `Apprenant \' ${apprenantDeleted.fullName} \' Deleted`, data: apprenantDeleted })
+                    })
                 })
                 .catch(err => {
-                    res.json({ message: 'we found somes issues', error: err.message })
+                    const message = 'Nous avons rencontre une erreur lors de la suppression de cet apprenant, veuillez reessayer plus tard...'
+                    res.status(500).json({ message })
                 })
-            })
-            .catch(err => {
-                res.json({ message: 'error', error: err.message })
-            })
         
             
     })
